Store character as an object
The snake (currently just the 'head') is now stored as an array of
objects, each of which contains a text character to be displayed, and
the row and column references to its position. Instead of hard coding
the 'o' and using global variables to track its position, the script
now updates the object in the array, which is more extensible.