Use JS to build the play area
Originally the table was hard-coded into the HTML, which made changing
the layout challenging (especially trying to do so on-the-fly). Now the
JS includes a DOMContentLoaded listener which sets up the structure of
the table, row by row and column by column, once the page is ready.
This change makes it very easy to adjust the dimensions manually, and
will allow implementation of a UI that lets the player decide how big a
play area they want.