Feed the snake
makeFood() creates a semicolon in a random cell of the table. The snake
can 'eat' the semicolon by passing over that cell (so overwriting the
contents).

Food is only created at the start of the game for now and has no effect
on the snake. There will need to be a check on whether the cell is empty
before extra food can be created mid-game.