Start the game with a key press instead of click
Since the snake is controlled with the keyboard and not the mouse, this
is more convenient. By pressing an arrow key the player can also set the
snake's initial movement direction.

Extracted code to start the game and to display body parts into separate
functions so they can be easily reused.