Update handling of direction changes
Since it is possible to enter multiple keystrokes in the 0.1 seconds
between move calls, the player could change direction e.g. from right to
up to left between calls, meaning the head never actually moves up but
goes from left on one call to right on the next and runs into the body.
The fast pace of updates makes it unreasonable for the player to observe
every update individually and determine when the change of direction has
taken place.

This issue was fixed by having any change in direction immediately call
move() without waiting for the regular timeout. To avoid any conflict
with the timeout, the first thing move() does now is to delete it if it
exists.