Fix bug causing snake to die when the game starts
If any key other than an arrow key was pressed, the game would start but
the snake would instantly die without moving. The cause was traced to
checkForDirectionChange() which compares old and new row/column change
values; the variables for recording new values (when an arrow key is
pressed) were initialised empty, and then immediately swapped with the
default direction vector because the method only checks whether the
direction is being reversed, and the 'undefined' values do not meet that
condition. This effectively set the direction vector to all zeroes and
caused the snake's head to crash into itself.

Also fixed two non-critical code errors, which had no impact on normal
gameplay but did not properly reflect the purpose of those lines.