Implement basic 'Comet' variant
Comet currently works in the same way as Snaky, but the body is
displayed as coloured squares which are shaded according to each
element's position in the tail.

The relevant changes are owned by the Comet constructor, so that the
same game controller can be used for both Snaky and Comet. Option
buttons will eventually be added to index.html to set the avatar style
based on the player's selection.