Update calculation of hues for Comet
RGB colours now used instead of hex for ease of dynamically calculating
the hue of each body part, based on both its position in the array and
the length of the array. Said calculation has been moved to the
addToCometBody method so that the hues are only recalculated when the
body changes, rather than every time the comet moves.

During this update I decided to change the style used to reference the
properties on a bodyPart; bodyPart['row'] is now bodyPart.row (likewise
for other properties) to better reflect that the components making up a
comet/snake body are JS objects in their own right.