Disintegrate comet tail on collisions
If the comet hits its own body, the cut off part of the tail will now
disappear one part at a time with a brief colour change, as though it is
being 'burned up' from the affected part.

When the comet hits a wall, the whole body disintegrates except for the
head.