Add functions that allow dynamic resizing of cells
By assigning a `style` element to a named variable and attaching it to
the document body, we can change the CSS at will by writing a new line
of innerHTML to the element.

For clarity, the style element is given a specific purpose (setting the
width & height of table cells) and the innerHTML is overwritten rather
than extended. Multiple style element variables can be declared if we
want to dynamically restyle other elements as well, e.g. base text size.