Dynamically resize gameView based on window height
The scaleGameView function reads the current height (in pixels) of the
page and the elements which will be visible when the game is live. Based
on these values, a suitable height is calculated for the rows in the
gameView table and this is used to call the setCellSize function, which
now uses px rather than em and also sets the font-size for consistency
of proportions at different scales.

An eventListener calls scaleGameView any time the window is resized; it
is also triggered when the zoom level changes (default behaviour). In
addition, createGameView calls scaleGameView after the table has been
set up, in case any parameters (e.g. numberOfRows) have changed.