Disable buttons while game is live
So that the game can be played in-browser on a touch screen device, the
buttons are now disabled when the game starts and re-enabled when the
game ends.

Touch screen users will eventually be able to control the comet/snake by
tapping near the top/right/bottom/left of the screen. Swiping would be
preferable but may conflict with browser forward/back commands, as some
devices recognise these gestures as browser actions by default.