Change direction based on screen taps
The direction is set to down, left, right or up based on whether the
screen was tapped near the bottom, left, top or right edge of the game
'window' - the boundaries of the snake/comet's movement.

Brief testing on a laptop with a touch screen suggest these anchors are
not ideal, as it is easy for a player to lose track of where their hands
are relative to the game window while they are tracking the movement of
the snake. It would be better if the tap position can be compared to the
current location of the snake's head, for instance.