Change direction on tapping based on head position
Tapping behaviour has been changed to calculate approximately where the
snake's head is, and switch direction based on the screen being tapped
above/below/left/right of that position.

To ensure every tap results in a change of direction, the current value
of rowChangeAmount is used to determine the appropriate axis i.e. if the
snake is already moving right/left, the direction will be changed based
on whether the screen was tapped above/below the snake's head; otherwise
the tap location will be considered as right or left of the head. This
is much easier for the player to determine while the snake is moving.

Initialisation of the row change and column change variables was updated
to better accommodate the above. By default, if the game is started by a
tap rather than a key press, the snake initially moves right.