Selectively disable user scaling on mobile devices
The player may need to tap the screen rapidly to turn sharp corners. By
default, this action commonly zooms in and out of the page, which is
undesirable while the game is in progress. However, scaling should be
allowed while the game is not live, so it is only deactivated when the
game starts and it is reinstated when the game ends.

Note: testing in Chrome on a Windows 10 laptop failed, but testing in
Chrome on a mobile running Android 8.1 was successful. Comments on
Stack Overflow indicate this discrepancy is common between mobile and
'non-mobile' devices.