Fix (?) bug where food sometimes wouldn't spawn
Occasionally after food was eaten, no new food would appear. The cause
is not obvious, but on reviewing the code it appears this phenomenon may
have been due to the sequence of events when the snake/comet body moves.

Each body part moves in turn, temporarily leaving a blank space in its
old position until the next part moves to occupy it. The check for food
is carried out for each body part, before it actually moves, including
creation of new food in an empty space.

It is possible that the sequence of checking and writing to cells is not
always executed in the intended order, so that a space is declared empty
immediately before both a body part and a piece of food try to occupy
it. Food created in such a space would be cleared and overwritten by a
body part within one round of movement, without being detected and so
without spawning more food.

Food handling is therefore now deferred to the end of the move cycle by
setting and checking a variable. IT IS NOT KNOWN WHETHER THIS FIXES THE
BUG as it was never replicable, but normal play testing after the change
showed expected behaviour.