Change layout and update styles
The page is now based on a widescreen landscape screen, with the banner
and controls on the right hand side. This allows the gameView table to
be made larger, and negates the need to disable eventListeners while the
instructions/settings panels are open.

This commit introduces elements which can be further manipulated to make
the page rearrange itself based on screen orientation and proportions,
comparing window width and height to determine the optimum display (a
planned future development).