Reduce default speed
Following feedback from users, the available game speeds are slightly
slower than they were before.

Speed is now determined at the start of the game rather than every time
the setting was changed, as there was no benefit to this.

Default speed is no longer hard coded - this was missed in a previous
update which changed the default radio button in the settings, but did
not change the default speed to match it.