Force snake to move right when the game starts
It was possible for the snake to die instantly on starting the game, for
example if the game was started by pressing the left arrow key. Various
options were considered to continue allowing the first keydown (or tap)
to also set the direction, but implementing this proved non-trivial and
removing the bug takes priority.