Give the snake a more snake-like body
The snake's body is now composed of 'wedges', each of which is oriented
every turn to 'follow' the body part in front of it.

Both the Comet and Snaky have had their digestFood methods updated to
display the newly created body part immediately, since it is now added
after the body has finished moving.