Fix game-breaking bugs in Edge
Critical bugs relating to screen orientation and FormData, which were
only affecting the MS browsers, have been trapped and compensated for
where possible.

Internet Explorer continues to raise numerous exceptions, so is now
formally not supported. It is anticipated that Edge will in general not
be supported either and some issues remain, but the game is currently
playable with reduced functionality, so occasional monitoring of its
performance in Edge may be of value to users.