Add code allowing Snaky to teleport through walls
The bodyPart which hit the wall is now passed through to the avatar's
hitWall method; Comet ignores it completely, but Snaky checks whether
teleporting is enabled and, if so, recalculates the co-ordinates to put
the bodyPart on the opposite side of the gameView, referring back to
tryToMove() to check for food/collisions etc in the new location.

Teleport is set to false for now with no option in-game to change it,
although it could be hacked via dev tools to set avatar.teleport = true.