Replace screen.orientation with CSS media query
Code for detecting orientation is now based on a simple comparison of
viewport width & height, which should have improved compatibility over
the HTML5 screen orientation API and is easier to work with.

Adjustments have not yet been made for 'barely portrait' aspect ratios
where the width is only slightly less than the height. This means the
current implementation works well on a smartphone, or in landscape mode
on a computer, but if the browser window on a computer is resized so as
to trigger the change to portrait mode the game can be started with the
instructions/settings open even though the play area is not visible.