Update appearance of page - MAJOR CHANGES
Although most of the HTML elements and their relative positioning are
basically the same, the way certain elements are housed and resized has
been changed to make use of more advanced CSS, particularly flexbox to
more easily rearrange the major segments on resizing or changing screen
orientation.

Settings and instructions now replace the play area when opened.

Changed the order in which some actions are executed to work better with
the style changes, namely when hiding/unhiding elements.